Fixes TypeError bug when you have only one dash
If the current dash happens to be the next dash, the iframe was
removed before the dashthe dash was sent on to be displayed again,
causing a TypeError when the .display method referred to the .iframe
property.

If you have a single dashboard, when testing for instance... this bug
was made evident.